Skala moved to MainForm and weight bug fixed!
